requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
tenacity>=8.2.0
tqdm>=4.65.0
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
